some mods to counter kicad lib cleaning
kicad/eeschema does modify the lib after loading it and removes complete fields when the value is empty, ie. "".
Thus, if the csv doesn't contain something we add an underscore "_"..
Then the fields added (for example datasheet URL or footprint 'links' are pretty big and in the way.. this script mods the textsize and visibility to make them pretty tiny.